All Room Lodging Process filled - no data labels
There is no data labels and connections for each transaction
process, the data labels and connection, I am not sure how
to present this part to a proper label and connection.

I am not sure what type of data will be present between the process.
Might have to ask what exactly goes in the process with the
group members.