Customer Availing Current Services Process
Finished the catering services and connecting the processes
for the catering process. But still no data label.

There is a part of the business process that was overlooked,
there is no business process for handling the stay or the
situation wherein the customers stays within the
business facilities.

Not part of the documentation. The check-out process
and what are the options of the customer for cancellation
or termination of the service.

For the rooms lodging the process is clear, the 50% non-refund
is clear once the room is paid for and the customer wants to
cancel or terminate the order. If there are cases wherein
the customer wants to negotiate the refund, I am not sure
of the exact process.

The step-8 of the documentation should have the contents for
the catering service to handle the post and during catering
service for the customer.