25% of Data Connection is labeled
Not sure if the labels are accurate and needs to be properly
reviewed. Viewing the .drawio file in github is not possible
so the file will have to be transferred to google drive for review.

The connections I can label are the ones that I
am sure of will represent data from one process to
another process.

Not sure how this will work with the catering
service.