Labels for Data Connection for Long Connections
I used two labels for connections that are long
and would span quite a large distance for easy tracking
of what information the line represents.

The room lodging and catering service have long Lines
of connection to the data storage, tracking down what
data the long lines represent would be a hard process
because there are many processes.

There is a gap between the content of the
documentation and the content of the data flow diagram.
I am not sure if the contents of the data flow diagram
is accurate to the documentation.