Started Placing Numbers for Process Orders
I want to follow the example placed in the
previous SAD projects, but I don't see any processes
that are used by multiple processes and processes
that can branch out to multiple processes.

So the part with the .1 and .2 and .[number]
are an improvisation that represents branching
of processes. Not sure if this is part of the
gane and sarson notation.

But the decimals are part of the gane and sarson notation
since the previous SAD project documentation had decimal
numbers in their DFD.

The numbers for the process orders is subject
to future changes and highly unstable.To be updated
and yet to be reviewed.