Added Processes on Catering Preparations
I did not add the proper processes for
catering services because I did not think
that this will be part of the
data flow diagram. So just to be safe I
added new processes.

The Process numbers are still partial and up-to
changes because I not sure if the data flow
diagram is still a level-zero data flow diagram.

Will be changed later, if the details are too
specific and needs to broader or simplified.