Finish All Labels, Data Connections & Processes
The recent processes for the catering and
management of general business maintenance and
operations, I'm not sure it the data processed
and sent to the management and owner is accurate.

Might need a review and revision.

The overall DFD is too big and needs to be simplified
to be printed or be broken up into different pieces
to be printed.