Enlarged Text Size, Reposition, & Revised Labels
The text size is too small when the exported
DFD is presented to the Document. I tried to Enlarge
the font size to 36, and repositioned some of the
processes to move towards the center and reduce overall
width of DFD.

Broke the positioning of the labels and connections
between the processes might need to be reworked later.