FINISH FULL DFD & DFD Components
The Data Flow Diagram is finished but still needs to be
organized and divided into components, the full data flow
diagram is too big and needs to be divided into smaller
components for easier presentation on the documentation file.

Though I did not have to create a different file for each of
the DFD component, I could have just used a layer
and change the visibility of the components,
which I did not even know that there are layers in
draw.io, just figured it out now as of writing this commit message.

The transactions, room lodging handling, and catering handling
are the three main sections that were divided and placed into
their own .drawio file. I did not export them with the vscode plugin
because the background would be black by default which is costly
in printing and bad in readability for small extremenly texts.

Changed the overall structure of the project to organize files
because things were getting cluttered.